WIP - specific constants moved to workflow constant folders
