+++ src/app/constants/index.ts
@@ -1,5 +1,4 @@
 import { ACCOUNT_ENQUIRY_DEFAULT_STATE, ACCOUNT_ENQUIRY_DEFAULT_STATE_SEARCH } from './account-enquiry-default-state';
-import { ACCOUNT_ENQUIRY_SEARCH_FORM_FIELD_ERRORS } from './account-enquiry-search-form-field-errors';
 import { API_PATHS } from './api-paths';
 import { ADDRESS_LINE_ONE_FIELD_ERRORS } from './common/address-line-one-field-errors';
 import { ADDRESS_LINE_THREE_FIELD_ERRORS } from './common/address-line-three-field-errors';
@@ -8,30 +7,18 @@ import { DATE_OF_BIRTH_FIELD_ERRORS } from './common/date-of-birth-field-errors'
 import { NATIONAL_INSURANCE_FIELD_ERRORS } from './common/national-insurance-field-errors';
 import { POST_CODE_FIELD_ERRORS } from './common/post-code-field-errors';
 import { DEFENDANT_TYPES_STATE } from './defendant-types-state';
-import { MANUAL_ACCOUNT_CREATION_ACCOUNT_DETAILS_FIELD_ERROR } from './manual-account-creation-account-details-field-errors';
 import { MANUAL_ACCOUNT_CREATION_ACCOUNT_DETAILS_STATE } from './manual-account-creation-account-details-state';
-import { MANUAL_ACCOUNT_CREATION_CONTACT_DETAILS_FIELD_ERROR } from './manual-account-creation-contact-details-field-errors';
 import { MANUAL_ACCOUNT_CREATION_CONTACT_DETAILS_STATE } from './manual-account-creation-contact-details-state';
-import { MANUAL_ACCOUNT_CREATION_NESTED_BUTTONS } from './manual-account-creation-nested-buttons';
-import { MANUAL_ACCOUNT_CREATION_NESTED_ROUTES } from './manual-account-creation-nested-routes';
-import { MANUAL_ACCOUNT_CREATION_EMPLOYER_DETAILS_FIELD_ERROR } from './manual-account-creation-employer-details-field-errors';
 import { MANUAL_ACCOUNT_CREATION_EMPLOYER_DETAILS_STATE } from './manual-account-creation-employer-details-state';
-import { MANUAL_ACCOUNT_CREATION_PARENT_GUARDIAN_DETAILS_FIELD_ERROR } from './manual-account-creation-parent-guardian-details-field-errors';
 import { MANUAL_ACCOUNT_CREATION_PARENT_GUARDIAN_DETAILS_STATE } from './manual-account-creation-parent-guardian-details-state';
-import { MANUAL_ACCOUNT_CREATION_PERSONAL_DETAILS_FIELD_ERROR } from './manual-account-creation-personal-detail-field-errors';
 import { MANUAL_ACCOUNT_CREATION_COURT_DETAILS_STATE } from './manual-account-creation-court-details-state';
 import { MANUAL_ACCOUNT_CREATION_PERSONAL_DETAILS_STATE } from './manual-account-creation-personal-details-state';
-import { MANUAL_ACCOUNT_CREATION_COMPANY_DETAILS_FIELD_ERROR } from './manual-account-creation-company-details-field-errors';
 import { MANUAL_ACCOUNT_CREATION_COMPANY_DETAILS_STATE } from './manual-account-creation-company-details-state';
 import { MANUAL_ACCOUNT_CREATION_BUSINESS_UNIT_STATE } from './manual-account-creation-business-unit-state';
 import { MANUAL_ACCOUNT_CREATION_STATE } from './manual-account-creation-state';
 import { ROUTE_PERMISSIONS } from './route-permissions';
 import { TITLE_DROPDOWN_OPTIONS } from './title-dropdown-options';
-import { MANUAL_ACCOUNT_CREATION_ACCOUNT_STATUS } from './manual-account-creation-account-status';
 import { CUSTOM_ADDRESS_FIELD_IDS } from './common/custom-address-field-ids';
-import { MANUAL_ACCOUNT_CREATION_PERSONAL_DETAILS_ALIAS } from './manual-account-creation-personal-details-alias';
-import { MANUAL_ACCOUNT_CREATION_COMPANY_DETAILS_ALIAS } from './manual-account-creation-company-details-alias';
-import { MANUAL_ACCOUNT_CREATION_COURT_DETAILS_FIELD_ERRORS } from './manual-account-creation-court-details-field-errors';
 
 export {
   ACCOUNT_ENQUIRY_DEFAULT_STATE,
@@ -40,19 +27,12 @@ export {
   API_PATHS,
   MANUAL_ACCOUNT_CREATION_EMPLOYER_DETAILS_STATE,
   MANUAL_ACCOUNT_CREATION_STATE,
-  ACCOUNT_ENQUIRY_SEARCH_FORM_FIELD_ERRORS,
-  MANUAL_ACCOUNT_CREATION_EMPLOYER_DETAILS_FIELD_ERROR,
-  MANUAL_ACCOUNT_CREATION_ACCOUNT_DETAILS_FIELD_ERROR,
   MANUAL_ACCOUNT_CREATION_ACCOUNT_DETAILS_STATE,
   DEFENDANT_TYPES_STATE,
   MANUAL_ACCOUNT_CREATION_CONTACT_DETAILS_STATE,
-  MANUAL_ACCOUNT_CREATION_CONTACT_DETAILS_FIELD_ERROR,
   MANUAL_ACCOUNT_CREATION_PARENT_GUARDIAN_DETAILS_STATE,
-  MANUAL_ACCOUNT_CREATION_PARENT_GUARDIAN_DETAILS_FIELD_ERROR,
   TITLE_DROPDOWN_OPTIONS,
-  MANUAL_ACCOUNT_CREATION_PERSONAL_DETAILS_FIELD_ERROR,
   MANUAL_ACCOUNT_CREATION_PERSONAL_DETAILS_STATE,
-  MANUAL_ACCOUNT_CREATION_COMPANY_DETAILS_FIELD_ERROR,
   MANUAL_ACCOUNT_CREATION_COMPANY_DETAILS_STATE,
   ADDRESS_LINE_ONE_FIELD_ERRORS,
   ADDRESS_LINE_TWO_FIELD_ERRORS,
@@ -60,13 +40,7 @@ export {
   POST_CODE_FIELD_ERRORS,
   DATE_OF_BIRTH_FIELD_ERRORS,
   NATIONAL_INSURANCE_FIELD_ERRORS,
-  MANUAL_ACCOUNT_CREATION_NESTED_BUTTONS,
-  MANUAL_ACCOUNT_CREATION_NESTED_ROUTES,
-  MANUAL_ACCOUNT_CREATION_ACCOUNT_STATUS,
   CUSTOM_ADDRESS_FIELD_IDS,
-  MANUAL_ACCOUNT_CREATION_PERSONAL_DETAILS_ALIAS,
-  MANUAL_ACCOUNT_CREATION_COMPANY_DETAILS_ALIAS,
   MANUAL_ACCOUNT_CREATION_COURT_DETAILS_STATE,
   MANUAL_ACCOUNT_CREATION_BUSINESS_UNIT_STATE,
-  MANUAL_ACCOUNT_CREATION_COURT_DETAILS_FIELD_ERRORS,
 };
