+++ src/app/pages/manual-account-creation/company-details/constants/manual-account-creation-company-details-alias.ts
@@ -1,6 +1,6 @@
 import { Validators } from '@angular/forms';
-import { IFormArrayControlValidation } from '../interfaces/form-array-control-validation.interface';
-import { alphabeticalTextValidator } from '../validators';
+import { IFormArrayControlValidation } from '@interfaces';
+import { alphabeticalTextValidator } from 'src/app/validators';
 
 export const MANUAL_ACCOUNT_CREATION_COMPANY_DETAILS_ALIAS: IFormArrayControlValidation[] = [
   {
