+++ src/app/pages/manual-account-creation/constants/manual-account-creation-nested-buttons.ts
