+++ src/app/pages/manual-account-creation/contact-details/constants/manual-account-creation-contact-details-field-errors.ts
