+++ src/app/pages/manual-account-creation/contact-details/contact-details-form/contact-details-form.component.ts
@@ -7,7 +7,6 @@ import {
   GovukErrorSummaryComponent,
   GovukTextInputComponent,
 } from '@components';
-import { MANUAL_ACCOUNT_CREATION_CONTACT_DETAILS_FIELD_ERROR, MANUAL_ACCOUNT_CREATION_NESTED_ROUTES } from '@constants';
 import { ManualAccountCreationRoutes } from '@enums';
 import { IFieldErrors, IManualAccountCreationContactDetailsForm } from '@interfaces';
 import {
@@ -15,6 +14,8 @@ import {
   optionalEmailAddressValidator,
   optionalPhoneNumberValidator,
 } from 'src/app/validators';
+import { MANUAL_ACCOUNT_CREATION_CONTACT_DETAILS_FIELD_ERROR } from '../constants/manual-account-creation-contact-details-field-errors';
+import { MANUAL_ACCOUNT_CREATION_NESTED_ROUTES } from '../../constants/manual-account-creation-nested-routes';
 
 @Component({
   selector: 'app-contact-details-form',
