+++ src/app/pages/manual-account-creation/contact-details/contact-details.component.ts
@@ -3,7 +3,7 @@ import { FormParentBaseComponent } from '@components';
 import { ManualAccountCreationRoutes } from '@enums';
 import { IManualAccountCreationContactDetailsForm } from '@interfaces';
 import { ContactDetailsFormComponent } from './contact-details-form/contact-details-form.component';
-import { MANUAL_ACCOUNT_CREATION_NESTED_ROUTES } from '@constants';
+import { MANUAL_ACCOUNT_CREATION_NESTED_ROUTES } from '../constants/manual-account-creation-nested-routes';
 
 @Component({
   selector: 'app-contact-details',
