+++ src/app/pages/manual-account-creation/court-details/court-details.component.ts
@@ -1,7 +1,6 @@
 import { ChangeDetectionStrategy, Component, inject } from '@angular/core';
 import { CourtDetailsFormComponent } from './court-details-form/court-details-form.component';
 import { FormParentBaseComponent } from '@components';
-import { MANUAL_ACCOUNT_CREATION_NESTED_ROUTES } from '@constants';
 import { ManualAccountCreationRoutes } from '@enums';
 import {
   IAutoCompleteItem,
@@ -14,6 +13,7 @@ import { CourtService, LocalJusticeAreaService } from '@services';
 import { Observable, forkJoin, map } from 'rxjs';
 import { CommonModule } from '@angular/common';
 import { RouterModule } from '@angular/router';
+import { MANUAL_ACCOUNT_CREATION_NESTED_ROUTES } from '../constants/manual-account-creation-nested-routes';
 
 @Component({
   selector: 'app-court-details',
