+++ src/app/pages/manual-account-creation/create-account/create-account-form/create-account-form.component.ts
@@ -10,10 +10,10 @@ import {
   GovukErrorSummaryComponent,
   AlphagovAccessibleAutocompleteComponent,
 } from '@components';
-import { MANUAL_ACCOUNT_CREATION_ACCOUNT_DETAILS_FIELD_ERROR } from '@constants';
 import { ManualAccountCreationRoutes, RoutingPaths } from '@enums';
 import { IAutoCompleteItem, IFieldErrors, IManualAccountCreationAccountDetailsState } from '@interfaces';
 import { DEFENDANT_TYPES_STATE } from 'src/app/constants/defendant-types-state';
+import { MANUAL_ACCOUNT_CREATION_CREATE_ACCOUNT_FIELD_ERROR } from '../constants/manual-account-creation-account-details-field-errors';
 
 @Component({
   selector: 'app-create-account-form',
@@ -43,7 +43,7 @@ export class CreateAccountFormComponent extends FormBaseComponent implements OnI
     ([key, value]) => ({ key, value }),
   );
 
-  override fieldErrors: IFieldErrors = MANUAL_ACCOUNT_CREATION_ACCOUNT_DETAILS_FIELD_ERROR;
+  override fieldErrors: IFieldErrors = MANUAL_ACCOUNT_CREATION_CREATE_ACCOUNT_FIELD_ERROR;
 
   /**
    * Sets up the employer details form with the necessary form controls.
