+++ src/app/pages/manual-account-creation/employer-details/employer-details-form/employer-details-form.component.ts
@@ -7,10 +7,6 @@ import {
   FormBaseComponent,
   GovukCancelLinkComponent,
 } from '@components';
-import {
-  MANUAL_ACCOUNT_CREATION_EMPLOYER_DETAILS_FIELD_ERROR,
-  MANUAL_ACCOUNT_CREATION_NESTED_ROUTES,
-} from '@constants';
 import { ManualAccountCreationRoutes } from '@enums';
 import { IFieldErrors, IManualAccountCreationEmployerDetailsForm } from '@interfaces';
 import {
@@ -19,6 +15,8 @@ import {
   optionalPhoneNumberValidator,
   specialCharactersValidator,
 } from 'src/app/validators';
+import { MANUAL_ACCOUNT_CREATION_EMPLOYER_DETAILS_FIELD_ERROR } from '../constants/manual-account-creation-employer-details-field-errors';
+import { MANUAL_ACCOUNT_CREATION_NESTED_ROUTES } from '../../constants/manual-account-creation-nested-routes';
 
 @Component({
   selector: 'app-employer-details-form',
