+++ src/app/pages/manual-account-creation/parent-guardian-details/parent-guardian-details-form/parent-guardian-details-form.component.ts
@@ -16,7 +16,6 @@ import {
   ADDRESS_LINE_TWO_FIELD_ERRORS,
   CUSTOM_ADDRESS_FIELD_IDS,
   DATE_OF_BIRTH_FIELD_ERRORS,
-  MANUAL_ACCOUNT_CREATION_PARENT_GUARDIAN_DETAILS_FIELD_ERROR,
   NATIONAL_INSURANCE_FIELD_ERRORS,
   POST_CODE_FIELD_ERRORS,
 } from '@constants';
@@ -30,6 +29,7 @@ import {
   specialCharactersValidator,
   alphabeticalTextValidator,
 } from 'src/app/validators';
+import { MANUAL_ACCOUNT_CREATION_PARENT_GUARDIAN_DETAILS_FIELD_ERROR } from '../constants/manual-account-creation-parent-guardian-details-field-errors';
 
 @Component({
   selector: 'app-parent-guardian-details-form',
