+++ src/app/pages/manual-account-creation/personal-details/personal-details-form/personal-details-form.component.ts
@@ -21,9 +21,6 @@ import {
   ADDRESS_LINE_TWO_FIELD_ERRORS,
   CUSTOM_ADDRESS_FIELD_IDS,
   DATE_OF_BIRTH_FIELD_ERRORS,
-  MANUAL_ACCOUNT_CREATION_NESTED_ROUTES,
-  MANUAL_ACCOUNT_CREATION_PERSONAL_DETAILS_ALIAS,
-  MANUAL_ACCOUNT_CREATION_PERSONAL_DETAILS_FIELD_ERROR,
   NATIONAL_INSURANCE_FIELD_ERRORS,
   POST_CODE_FIELD_ERRORS,
   TITLE_DROPDOWN_OPTIONS,
@@ -39,6 +36,9 @@ import {
   optionalValidDateValidator,
   specialCharactersValidator,
 } from 'src/app/validators';
+import { MANUAL_ACCOUNT_CREATION_PERSONAL_DETAILS_FIELD_ERROR } from '../constants/manual-account-creation-personal-detail-field-errors';
+import { MANUAL_ACCOUNT_CREATION_PERSONAL_DETAILS_ALIAS } from '../constants/manual-account-creation-personal-details-alias';
+import { MANUAL_ACCOUNT_CREATION_NESTED_ROUTES } from '../../constants/manual-account-creation-nested-routes';
 
 @Component({
   selector: 'app-personal-details-form',
